ServerSentEvents/onopen-onmessage-onerror-support.html

<!DOCTYPE html>
<html>
<head>
	<title>The event handlers (onopen, onmessage, onerror) must be supported, as IDL attributes, by all objects implementing the EventSource interface:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The event handlers (onopen, onmessage, onerror) must be supported, as IDL attributes, by all objects implementing the EventSource interface:</h1>
	<div id="log"></div>
  <script>
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0025');
    
    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.onmessage = function(e){
    //alert(dat);
    };
    	
    s.onerror = function(e){
    	/* Connection closed. */
    };
    
    test(function(){assert_true(typeof(s.onopen) == "function")},"The onopen event handler must be supported.");
    test(function(){assert_true(typeof(s.onmessage) == "function")},"The onmessage event handler must be supported.");
    test(function(){assert_true(typeof(s.onerror) == "function")},"The onerror event handler must be supported.");
    
  </script>
  </div>
</body>
</html>
